+ pixelup: update: upgrade both slots for bootloader update
we must to both slots because it has a non-reversible efuse it seems,
when it writes, it cannot go back, and won't run on the older bootloader
anymore.  so if upgraded slot fails to boot, then old slot is tried, but
fuse has been forward-set, neither will be able to boot and phone will
be brick.

this is specifically mentioned in google factory image download page,
worthy to add the stuff here.  what we did was just sideload again after
upgrade, apparently this updates the inactive slot and then sets it
active.  no reason to do this routinely (update both slots) because it's
fine for the non-working slot to have the old version, as long as the
bootloader loads, it can still be used to flash other images and fix
problems, etc.  this is why we have to do it in the case that bootloader
is updated for major android version upgrade (as in this case, from
android 12 to 13)