- pixelup: make clear no need to do both slots unless major ver up
these extra instructions to do the other slot are because there is no
way even to flash anything if the bootloader gets upgraded on one side
but the irreversible ROM flash gets set to later android version.  if
original slot breaks no way to ever fix the phone.  but this is only an
issue when doing major version upgrade usually and not necessary at
other times, since the phone will still boot the bootloader even on the
old/stale slot normally