- py3build: py: [ug]id with tar so we can change pip interp in relocate
we have to change the pip interpreter before creating the archive, so it
gets extracted correctly, so that section of docs needs to be moved up.

this means however that we would have to edit the files as root, since
we necessarily have chowned to root at the end of the section that
describes the relocate root.  so instead of doing a chown at all, we
specify root user/group to tar, so pip interpreter change can come first
and doesn't require modifying files that have been root-owned.  this is
probably better anyways as all the files can be user-owned without ever
needing to use sudo at all and eliminating one step.