^ py3build: dynlib: don't mind some libs that are never present
when determining the list of shared object dependencies of the dynlibs,
we ignore ones that don't change across releases.  the libs this patch
removes from consideration actually never show up for any dynload on
u18+ systems.  these ones must have been old ones on u14 and/or old
centos, which we are not building for anymore.