+ pixelup: add alternate backup method that avoids a staged copy
previously, we had to stage the whole dir and zip it on our system, but
this required that we have like 80G of space.  if we pull it with tar
directly, it never needs to hit the ssd chips until it's already in its
final form.  we leave the gzip in for now in case the bottleneck is the
cable speed.  we only got 28MB/s copying the backup but it was not using
our thunderbolt cable, so we'll test whether the gzip helps or hurts
next time we update.