- pixelup: update image to march 2025, no magisk update this month
everything went fine, although had to change Nova setting to obey dark
mode from System Setting, it was set for light mode.  not clear how this
happened but others on the Internet have complained about the same
thing, which is where we got the idea to check the dark mode setting.