- pixelup: switch to non-compressed tar and multithread zstd on host
we have fast thunderbolt, the backup was only going 25M/s average and
unclear what bottleneck was but thinking it was the initial compression
and single thread zstd.  we should not need to compress twice because
thunderbolt should be pretty good and keep up.  will experiment with
this each month and see the fastest method.  downside is it will make
our laptop very slow since all processors will be saturated, but we can
adjust that as needed depending on our needs at the time.